single table for property & post
