+++ mkanak_spring/src/main/java/mkanak_spring/model/Apartment.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor
 @Table(name = "apartment")
-public class Apartment extends Property {
+public class Apartment extends Post {
     @Column(name = "apartment_level")
     private int level;
     @Column(name = "has_elevator")
