+++ mkanak_spring/src/main/java/mkanak_spring/model/PostManager.java
@@ -1,24 +1,21 @@
 package mkanak_spring.model;
 
-import mkanak_spring.model.dao.ApartmentRepo;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class PostManager {
-
     public List<Post> getPersonPosts(int targetUserID, ViewingPreference preferences) {
         return null;
     }
     public List<Post> getHomePage(ViewingPreference preferences){
         List<Post> homePageContent= new ArrayList<>();
-        for(int i=0;i<5;i++){
-           // homePageContent.add(getDummyPost());
+        for(int i=0;i<5;i++) {
+            // homePageContent.add(getDummyPost());
         }
         return homePageContent;
     }
@@ -39,28 +36,8 @@ public class PostManager {
         return false;
     }
 
-    public Post addPost(Property property) throws ParseException {
-        Post newPost = new Post();
-    //    int a = (int) postDetails.get("area");
-     //   System.out.println("ADD: " + a);
-      //  JSONParser parser = new JSONParser();
-      //  JSONObject post = (JSONObject) parser.parse(postDetails.toString());
-     /*   if(post.get("type") == "villa") {
-            Villa property = new Villa();
-            property = buildVilla(post);
-        }
-        else {
-            Apartment property = new Apartment();
-            property = buildApartment(post);
-        }
 
-      */
-      //  newPost.setProperty(property);
-        /**newPost.setPublishDate();*/
-        return new Post();
-    }
-
-    public Property buildProperty(Property property, JSONObject post) {
+    public Post buildPost(Post property, JSONObject post) {
         property.setAddress((String) post.get("address"));
         property.setType((String) post.get("type"));
         property.setArea((Integer) post.get("area"));
@@ -76,7 +53,7 @@ public class PostManager {
 
     public Apartment buildApartment(JSONObject property) {
         Apartment apartment = new Apartment();
-        buildProperty(apartment, property);
+        buildPost(apartment, property);
         apartment.setLevel((Integer) property.get("level"));
         apartment.setElevator((boolean) property.get("elevator"));
         apartment.setStudentHousing((boolean) property.get("studentHousing"));
@@ -85,7 +62,7 @@ public class PostManager {
 
     public Villa buildVilla(JSONObject property) {
         Villa villa = new Villa();
-        buildProperty(villa, property);
+        buildPost(villa, property);
         villa.setNumberOfLevels((Integer) property.get("level"));
         villa.setHasGarden((boolean) property.get("hasGarden"));
         villa.setHasPool((boolean) property.get("hasPool"));
