+++ mkanak_spring/src/main/java/mkanak_spring/model/Villa.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor
 @Table(name = "villa")
-public class Villa extends Property {
+public class Villa extends Post {
    @Column(name = "has_garden")
    boolean hasGarden;
    @Column(name = "has_pool")
