--- mkanak_spring/src/main/java/mkanak_spring/model/dao/PropertyRepo.java
@@ -1,10 +1,0 @@
-package mkanak_spring.model.dao;
-
-import mkanak_spring.model.Property;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PropertyRepo extends JpaRepository<Property, Long> {
-}
