+++ mkanak_spring/src/main/java/mkanak_spring/model/services/PostServiceImpl.java
@@ -4,6 +4,7 @@ import mkanak_spring.model.*;
 import mkanak_spring.model.dao.PostDAO;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,32 +20,29 @@ public class PostServiceImpl implements PostService{
     public void savePost(JSONObject post) throws ParseException {
         PostManager postManager = new PostManager();
         System.out.println("Type: " + post.get("type"));
+        JSONParser parser = new JSONParser();
+        JSONObject object = (JSONObject) parser.parse(post.toString());
+        JSONArray pictures = (JSONArray) object.get("pictures");
         if(post.get("type").toString().compareTo("villa") == 0) {
             Villa property = new Villa();
             property = postManager.buildVilla(post);
-            Post newPost = postManager.addPost(property);
-            postDAO.savePost(newPost);
-            List<PropertyPicture> pictureList = postManager.buildPropertyPictures(post, newPost.getPostID());
-            property.setHasPictures(pictureList.size() != 0);
-            property.setPost(newPost);
+            property.setHasPictures(pictures.size() != 0);
             postDAO.saveVilla(property);
+            List<PropertyPicture> pictureList = postManager.buildPropertyPictures(post, property.getPropertyID());
             postDAO.saveAllPictures(pictureList);
         }
         else {
             Apartment property = new Apartment();
             property = postManager.buildApartment(post);
-            Post newPost = postManager.addPost(property);
-            postDAO.savePost(newPost);
-            List<PropertyPicture> pictureList = postManager.buildPropertyPictures(post, newPost.getPostID());
-            property.setHasPictures(pictureList.size() != 0);
-            property.setPost(newPost);
+            property.setHasPictures(pictures.size() != 0);
             postDAO.saveApartment(property);
+            List<PropertyPicture> pictureList = postManager.buildPropertyPictures(post, property.getPropertyID());
             postDAO.saveAllPictures(pictureList);
         }
     }
 
     @Override
     public List<Post> getAllPosts() {
-        return postDAO.getAll();
+        return postDAO.getAllPosts();
     }
 }
